requirements.txt
certifi==2018.11.29
chardet==3.0.4
Flask==0.10.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
opencv-python==3.4.5.20
Pillow==5.4.0
requests==2.21.0
urllib3==1.24.1
Werkzeug==0.14.1
